fix: address code review issues from PR #2
- Fix device_type double-quoting bug: strip JSON quotes before storing
  in SQLite so read path doesn't produce invalid JSON (silent data
  corruption where all device types fell back to Unknown)
- Fix logging guideline violations: repeat structured fields in message
  text across packet_capture_pnet, packet_capture_noop, and discovery
- Remove dead lookup_host call in hostname_resolver that fired a useless
  DNS query on every resolution
- Update hostname_resolver doc comments to match implementation
- Move inline config tests to separate file per project convention
- Clarify AGENTS.md logging rule: simple status messages without
  structured fields are fine as-is
