+++ source/daemon/crates/wardnetd/src/tests/mod.rs
@@ -1,3 +1,4 @@
+mod config;
 mod event;
 mod keys;
 mod oui;
